Andor: Avoid VC++ warning
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14368 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Andor/Andor.cpp
@@ -1573,7 +1573,7 @@ int AndorCamera::GetListOfAvailableCameras()
       if(DRV_SUCCESS!=ret)
         return ret;
       fReadOutTime = fKinetic;
-      SetExposureTime(currentExpMS_/1000.0f);
+      SetExposureTime(static_cast<float>(currentExpMS_ / 1000.0));
 
       
       ret = GetAcquisitionTimings(&fExposure,&fAccumTime,&fKinetic);
